First part of port of memory management to Rust (#347)
* initial commit

* updates to get tests passing

* fix memory allocation


* fix format

* Updates to example program


* Fix hang in example


* fix clippy


* Rename Copy Byte Array


* Additional changes for mem management


* Remove unused code/ rename/ refactor


* attempt to fix rsut fmt issue with comments


* Fix comments to prevent Rust fmt falures


* fix test options


* Update src/hyperlight_host/src/mem/layout.rs

fix typo in comments


* Update src/hyperlight_host/src/mem/layout.rs

fix typo in comments


* Update src/hyperlight_host/src/mem/guest_mem.rs

review suggestion


* Update src/hyperlight_host/src/mem/layout.rs

Fix type in comment


* Update src/hyperlight_host/tests_capi/guest_mem.c

make length a calc instead of constant


* Add context to error message


* fix pragma restore


* fix type in change

+++ src/Hyperlight/Hypervisors/HyperV.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using Hyperlight.Core;
 using Hyperlight.Native;
+using Hyperlight.Wrapper;
 
 
 namespace Hyperlight.Hypervisors
@@ -21,7 +22,7 @@ namespace Hyperlight.Hypervisors
         readonly bool virtualProcessorCreated;
         readonly ulong size;
 
-        internal HyperV(IntPtr sourceAddress, int pml4_addr, ulong size, ulong entryPoint, ulong rsp, Action<ushort, byte> outb, Action handleMemoryAccess) : base(sourceAddress, entryPoint, rsp, outb, handleMemoryAccess)
+        internal HyperV(IntPtr sourceAddress, ulong pml4_addr, ulong size, ulong entryPoint, ulong rsp, Action<ushort, byte> outb, Action handleMemoryAccess) : base(sourceAddress, entryPoint, rsp, outb, handleMemoryAccess)
         {
             this.size = size;
             WindowsHypervisorPlatform.WHvCreatePartition(out hPartition);
@@ -41,7 +42,7 @@ namespace Hyperlight.Hypervisors
                 registerValuesList.Add(new WindowsHypervisorPlatform.MyUInt128() { low = registerValueLow, high = registerValueHigh });
             }
 
-            AddRegister(WindowsHypervisorPlatform.WHV_REGISTER_NAME.WHvX64RegisterCr3, (ulong)pml4_addr, 0);
+            AddRegister(WindowsHypervisorPlatform.WHV_REGISTER_NAME.WHvX64RegisterCr3, pml4_addr, 0);
             AddRegister(WindowsHypervisorPlatform.WHV_REGISTER_NAME.WHvX64RegisterCr4, X64.CR4_PAE | X64.CR4_OSFXSR | X64.CR4_OSXMMEXCPT, 0);
             AddRegister(WindowsHypervisorPlatform.WHV_REGISTER_NAME.WHvX64RegisterCr0, X64.CR0_PE | X64.CR0_MP | X64.CR0_ET | X64.CR0_NE | X64.CR0_WP | X64.CR0_AM | X64.CR0_PG, 0);
             AddRegister(WindowsHypervisorPlatform.WHV_REGISTER_NAME.WHvX64RegisterEfer, X64.EFER_LME | X64.EFER_LMA, 0);
